Add `Direction` type and `perform_move()`
I also turned on a bunch of new Clippy warnings, and that led top the introduction of several instances of `#[must_use]` and `const`, and replacing type names with `Self` in impl blocks.

We also derived both `Copy` and `Clone` for `Pos` so that we didn't have to worry about things like ownership and references to `Pos` values. in functions like `current_pos()`.